Set charset to utf-8 to avoid warning. (Fixes #7)
Fixes warning in Firefox Browser Console:
"The character encoding of the HTML document was not declared. The
document will render with garbled text in some browser configurations if
the document contains characters from outside the US-ASCII range. The
character encoding of the page must be declared in the document or in
the transfer protocol."